Centered footer items (fixes #3), better Makefile
The new Makefile uses Bash bracket expansion to match all the unwanted
file types in the base project directory as well as the sections and exp
directories. This requires specifying that `/bin/bash` is the SHELL that
`make` should use for the commands. Certain non-Ubuntu-like Linux
distros may not be compatible with this setting.