Center of footer is actually centered, fixes #3
The previous "fix" for #3 (3339c41871d7) used a simple approach of
putting all three parts of the footer in the center footer `\cfoot` with
`\hfill`s in between values to achieve centering of text. The problem
was that if the footer name is wider than the page number and all three
fields are shown.

This fix uses all three footer fields (`\lfoot`, `\cfoot`, `\rfoot`),
which means that the contents of `\cfoot` are actually centered on the
page now, rather than just having equal amounts of whitespace on either
side of it.