New rExperience environment, fixes issue #8
The normal command for a new environment (\newenvironment) wasn't
verbose enough to allow for the body of the environment to be switched
between showing up (normal experience printing) and not (experience
summary, when the title-only flag is on).

So, the rExperience environment uses the \NewEnviron command from the
environ package. It has some pretty sweet features.