Improved spacing around pubs list, fixes issue #7
Removed some of the space appearing after the publication subsection
title (before the first entry), as well as after the end of the last
entry in the publications list.

Since there didn't appear to be any commands native to the biblatex
package to control this spacing, this is currently done with the
`\vspace{}` command, using `\parskip` as the unit of length. While the
results improve on the previous spacing, it is only an approximate
adjustment, mimicing the spacing around other comparable sections. It is
likely not worth the effort to get this to be exact, especially
considering the number of hours that went into commit 8675844 (it was a
lot...) to get things right.