rSection param to toggle titleOnly flag fixes #9
Like the rest of the document, the global `titleOnly` flag is off by
default. With this new feature, this default behavior remains unchanged,
so if the user doesn't specify a flag, `titleOnly` will continue to be
false. The user may manually set the flag to false with the new
parameter, but that doesn't seem practical.

There are two ways to set the flag: with and without explicitly stating
it should be true. The following two examples are equivalent:
```
#!latex

\begin{rSection}[titleOnly=true]{Section Name}
\end{rSection}
```
```
#!latex

\begin{rSection}[titleOnly]{Section Name}
\end{rSection}
```