A few improvements to bibliography
When printing the author names, BibLaTeX will now make the name bold if
the last name matches what is in the `\MyLastName` macro. This makes it
unnecessary to add `\textbf{}` to the `.bib` file directly, which makes
things easier on citation managers.

Also, fixed some spacing issues that only happen when attempting to
print a bibliography with no entries. What was happening before was a
`\vspace{}` with a negative amount was being printed regardless of
whether the actual bibliography was printed. So the next bibliography
ended up having extra negative `\vspace{}` commands right before it.