Re #7 standardized more spacing, fixed icon bug
While working on commit 0040de09, I noticed there were a couple places
where it would make sense to replace a hard-coded, negative `\vspace{}`
with a macro so that users can customize the spacing like in other
locations.

Also, for some reason the wrapper around the icon graphic was hard-coded
to a value that makes sense for my personal use, but isn't ideal for a
general use case. Changed this to set the width to be the same as the
icon's width.