Allow specifying field order, holds #12
Order of fields can now be specified using the `\setExpOrder{}` command
and one of three options:

* `tpnl`
* `nltp`
* `nptl`

Details about what these mean are given in the README. Not closing the
issue yet because I'm hoping there's a better way to specify order
options other than enumerating all 24 (4!) combinations by hand.

Also, introduced the `\expExtra{}` command for content that should
appear after the four fields, re #14. This was necessary to allow the
extra content to be displayed at the end of the header regardless of the
field order specified.