SELECTABLE FONTS!! re #4
This was a tricky option to support. Some of the things that I had to
use are: Experimental commands (that come from LaTeX3),
dynamically-loaded packages (which cannot be done when processing
options), using a `switch`/`case` block that compares strings with a
value in a variable (not as easy as you might think).

A key turning point in getting this to work was one particular post on
Stack Exchange:
http://tex.stackexchange.com/questions/314302/good-examples-of-tl-casenntf-and-str-casenntf