Fixed two bugs: bib title, Resume footers
In commit 8675844, one change I made was to start using BibLaTeX's
mechanism for showing the title for a set of publications. The problem I
didn't realize at the time was if the `title` option is given an empty
string, it still falls back to using the default bibliography heading.

The reason this is a problem is that the Resume document displays the
publications heading at a different hierarchical level than in the CV
document, which is why the Resume passes an empty string as the title
for the publications (which used to result in a title not displaying at
all). The end result was the Resume showed its own title and then
BibLaTeX showed its default title as well. Reverting back to the old
code helped this, and I added some spacing patches as well.

The second bug also affected Resume documents. A few circumstances made
the CV mode switcher (re #4, 92e307d) incompatible with how the footer
fields were populated. Specifically, the internal flags (toggles) that
indicated that certain content would or wouldn't be displayed got messed
up. This was partially because when `xkeyval` processes the keyword
options, it actually executes all the default options before processing
the user-supplied options. So, some flags were explicitly being set to
True and then False. A few extra checks (plus some minor code
improvements) fixed all this.