No more `ifthen` package commands, closes #16
All conditional checks now use commands from the `etoolbox` package. One
challenge I faced was getting the last page number as a number, which is
required by the `etoolbox` commands. The `pageslts` package uses a few
counters to keep track of the pages, and the one used in these revisions
(`pagesLTS.pagenr`) is actually the total number of pages from all
different page numbering styles. Although this is a bit more verbose
than this type of document needs, I decided to use it in case I refer
back to this implementation for a different document, hoping this would
be the most general use case.