Class options: icon and iconsize, re #4
One issue I had with this was that the errors don't actually show up in
the log. The full stop definitely takes place as it should, but the .log
file doesn't contain the custom errors I wrote.

Section 4.9 of this document discusses the `\ClassError` command:
http://mirrors.concertpass.com/tex-archive/macros/latex/doc/clsguide.pdf

I tried using different packages that are wrappers for the native error
commands, but they didn't seem to help. It's possible that LaTeX3's
coffin code layer might do the trick, but it also seems complicated.
They are documented in Part XVII (~pg 177) of this document:
http://ctan.mackichan.com/macros/latex/contrib/l3kernel/source3.pdf