Fixed bug from commit f7156dce, improved Makefile
It seems that the check I introduced in f7156dce is only for packages,
not arbitrary files. So since I was using it to check if a figure
actually exists before trying to use it as the icon, compiling generated
an error because of the check when it otherwise would compile just fine.

Also, the Makefile can now compile the document. Just specify the
target `cv` or `resume` and it will do the rest. (It does this via the
`latexmk` command, so that and `make` must be installed for this to
work.)